Fix and improve Apache FilesMatch rule.
The original regex "\.php" was inside of a Ruby string. Ruby would
parse the backslash leaving only the dot printed. Added a second
backslash that escapes the original one. This way, "\.php" is
actually printed making the regex correct.

In addition added another dot to the beginning of the regex. This
makes it match only files that has “.php” extension and not a file
named “.php” itself. It still matches files like “..php” though.